Add files via upload
--> Mudanças sujeitas a serem revertidas <--
Classe Maquina separada em:
- Exibidor
- ControleProduto
- ControleApurado

Classe Menu separada em:
- ControleEstoque

+++ include/Bebida.h
@@ -13,7 +13,7 @@ class Bebida : public Produto
         virtual ~Bebida();
 
         // setters
-        void setVolume(double vol);
+        void setVolume(int vol);
 
         // getters
         int getVolume();
